feat: add modular AI service
